// tests/src/types/dca/response/get_events.d.ts
export type EventData =
  | {
      dca_vault_funds_deposited: {
        amount: Coin;
      };
    }
  | {
      dca_vault_execution_triggered: {
        asset_price: Decimal;
        base_denom: string;
        quote_denom: string;
      };
    }
  | {
      dca_vault_execution_completed: {
        fee: Coin;
        received: Coin;
        sent: Coin;
      };
    }
  | {
      simulated_dca_vault_execution_completed: {
        fee: Coin;
        received: Coin;
        sent: Coin;
      };
    }
  | {
      dca_vault_execution_skipped: {
        reason: ExecutionSkippedReason;
      };
    }
  | {
      simulated_dca_vault_execution_skipped: {
        reason: ExecutionSkippedReason;
      };
    }
  | {
      dca_vault_cancelled: {};
    }
  | {
      dca_vault_escrow_disbursed: {
        amount_disbursed: Coin;
        performance_fee: Coin;
      };
    }
  | {
      dca_vault_post_execution_action_failed: {
        funds: Coin[];
        msg: SubMsgFor_Empty;
      };
    };
// ...
export type Uint128 = string;
/**
 * A fixed-point decimal value with 18 fractional digits, i.e. Decimal(1_000_000_000_000_000_000) == 1.0
 *
 * The greatest possible value that can be represented is 340282366920938463463.374607431768211455 (which is (2^128 - 1) / 10^18)
 */
export type Decimal = string;
export type ExecutionSkippedReason =
  | ("slippage_tolerance_exceeded" | "swap_amount_adjusted_to_zero")
  | {
      price_threshold_exceeded: {
        price: Decimal;
      };
    };
export type CosmosMsgFor_Empty =
  | {
      bank: BankMsg;
    }
  | {
      custom: Empty;
    }
  | {
      staking: StakingMsg;
    }
  | {
      distribution: DistributionMsg;
    }
  | {
      stargate: {
        type_url: string;
        value: Binary;
        [k: string]: unknown;
      };
    }
  | {
      ibc: IbcMsg;
    }
  | {
      wasm: WasmMsg;
    }
  | {
      gov: GovMsg;
    };
// ...
export type ReplyOn = "always" | "error" | "success" | "never";

export interface EventsResponse {
  events: Event[];
}
export interface Event {
  block_height: number;
  data: EventData;
  id: number;
  resource_id: Uint128;
  timestamp: Timestamp;
}
export interface Coin {
  amount: Uint128;
  denom: string;
  [k: string]: unknown;
}
// ...
export interface SubMsgFor_Empty {
  /**
   * Gas limit measured in [Cosmos SDK gas](https://github.com/CosmWasm/cosmwasm/blob/main/docs/GAS.md).
   */
  gas_limit?: number | null;
  /**
   * An arbitrary ID chosen by the contract. This is typically used to match `Reply`s in the `reply` entry point to the submessage.
   */
  id: number;
  msg: CosmosMsgFor_Empty;
  reply_on: ReplyOn;
  [k: string]: unknown;
}