// tests/src/types/dca/response/get_vault.d.ts
export type Uint128 = string;
// ...
export type Timestamp = Uint64;
// ...
export type Uint64 = string;
// ...
export type Addr = string;
/**
 * A fixed-point decimal value with 18 fractional digits, i.e. Decimal(1_000_000_000_000_000_000) == 1.0
 *
 * The greatest possible value that can be represented is 340282366920938463463.374607431768211455 (which is (2^128 - 1) / 10^18)
 */
export type Decimal = string;
/**
 * Binary is a wrapper around Vec<u8> to add base64 de/serialization with serde. It also adds some helper methods to help encode inline.
 *
 * This is only needed as serde-json-{core,wasm} has a horrible encoding for Vec<u8>. See also <https://github.com/CosmWasm/cosmwasm/blob/main/docs/MESSAGE_TYPES.md>.
 */
export type Binary = string;
export type PerformanceAssessmentStrategy = {
  compare_to_standard_dca: {
    received_amount: Coin;
    swapped_amount: Coin;
  };
};
export type VaultStatus = "scheduled" | "active" | "inactive" | "cancelled";
export type SwapAdjustmentStrategy =
  | {
      risk_weighted_average: {
        base_denom: BaseDenom;
        model_id: number;
        position_type: PositionType;
      };
    }
  | {
      weighted_scale: {
        base_receive_amount: Uint128;
        increase_only: boolean;
        multiplier: Decimal;
      };
    };
export type BaseDenom = "bitcoin";
export type PositionType = "enter" | "exit";
export type TimeInterval =
  | (
      | "every_block"
      | "every_minute"
      | "half_hourly"
      | "hourly"
      | "half_daily"
      | "daily"
      | "weekly"
      | "fortnightly"
      | "monthly"
    )
  | {
      custom: {
        seconds: number;
      };
    };
export type TriggerConfiguration =
  | {
      time: {
        target_time: Timestamp;
      };
    }
  | {
      price: {
        order_idx: Uint128;
        target_price: Decimal;
      };
    };

export interface VaultResponse {
  vault: Vault;
}
export interface Vault {
  balance: Coin;
  created_at: Timestamp;
  deposited_amount: Coin;
  destinations: Destination[];
  escrow_level: Decimal;
  escrowed_amount: Coin;
  id: Uint128;
  label?: string | null;
  minimum_receive_amount?: Uint128 | null;
  owner: Addr;
  performance_assessment_strategy?: PerformanceAssessmentStrategy | null;
  received_amount: Coin;
  slippage_tolerance: Decimal;
  started_at?: Timestamp | null;
  status: VaultStatus;
  swap_adjustment_strategy?: SwapAdjustmentStrategy | null;
  swap_amount: Uint128;
  swapped_amount: Coin;
  target_denom: string;
  time_interval: TimeInterval;
  trigger?: TriggerConfiguration | null;
}
export interface Coin {
  amount: Uint128;
  denom: string;
  [k: string]: unknown;
}
export interface Destination {
  address: Addr;
  allocation: Decimal;
  msg?: Binary | null;
}